BAH-3521 | Move custom address field extension to bahmni_address module
